assignment 3 part 3 style.css sections 1 to 3
First of all, download a couple of free-to-use fonts. Because this is a college, the fonts should be chosen to give the page a fairly serious, formal, trustworthy feel: a serif site-wide font for the general body text, coupled with a sans-serif or slab serif for the headings might be nice.
Use a suitable service to generate bulletproof @font-face code for these two fonts.
Apply your body font to the whole page, and your heading font to your headings.