p3 style.css section: links
Make it so that links are underlined by default, but when you hover or focus them the underline disappears.
Remove the default focus outline from ALL the links on the page.
Give the active state a noticeably different styling so it stands out nicely, but make it still fit in with the overall page design.
Make it so that external links have the external link icon inserted next to them.

id = "icon" were also added to the anchor tags for the external links