p3 style.css section: Lists
Make sure the spacing of your lists and list items works well with the styling of the overall page. Each list item should have the same line-height as a paragraph line, and each list should have the same spacing at its top and bottom as you have between paragraphs.
Give your list items a nice bullet appropriate for the design of the page. It is up to you whether you choose a custom bullet image or something else.